section 8 (partial): split hfs_fsck Phase 2 (btree) into submodule
Extract catalog B-tree structure checks + repairs into
src/fs/hfs_fsck/btree.rs (1322 lines): check_btree_structure,
check_map_nodes, check_node_bitmap_consistency, compare_catalog_keys,
check_key_ordering, check_offset_table, check_index_record_lengths,
check_leaf_record_lengths, walk_leaf_chain, check_index_sibling_links,
collect_referenced_nodes, plus repairs repair_btree_structure_leaves_only,
repair_map_nodes, repair_node_bitmap, repair_key_ordering,
fix_leaf_record_count, is_valid_leaf_node, extract_first_key,
rebuild_leaf_chain, rebuild_index_nodes, count_leaf_records.

The shared btree primitives record_key, extract_child_pointers, and
HFS_MAX_NRECS move with the module; extents.rs now imports them via
super::btree::... rebuild_index_nodes is re-exported pub(crate) from
mod.rs so hfs.rs callers (super::hfs_fsck::rebuild_index_nodes) keep
working unchanged.

hfs_fsck/mod.rs drops from 4353 to 3054 lines. Only Phase 3 (catalog
consistency) remains inside mod.rs. All 61 fsck tests pass.
