+++ docs/codecleanup.md
@@ -177,8 +177,8 @@ Conventions:
 - [ ] Catalog walking / record-range helpers shared with HFS+ (§1).
 - [ ] Audit debug `eprintln!` instances flagged in `hfs.rs:3068–3124` and replace with `log::debug!`.
 - [~] `hfs_fsck.rs` split started: extents-overflow code moved to its own submodule.
-  - **Done:** `src/fs/hfs_fsck.rs` is now `src/fs/hfs_fsck/mod.rs` (4353 lines, was 6588) plus three submodules: `extents.rs` (1334 lines, the 17 `*_extents_*` functions for the extents-overflow B-tree — `check_extents_btree_structure` / `repair_extents_btree_structure` + helpers + tests), `mdb.rs` (260 lines, Phase 1 MDB sanity — `validate_hfs_name`, `check_alternate_mdb`, `check_embedded_hfs_plus`, `check_mdb`), and `bitmap.rs` (698 lines, Phase 4 extent/allocation-bitmap cross-check + repairs — `check_extents_and_bitmap`, `collect_overflow_extents`, `mark_extents`, `repair_extents`, `repair_mdb_counts_and_bitmap`, plus the local `MAX_BITMAP_MISMATCHES` const and `hfs_debug_issue` helper). Helpers `HfsFsckCode`, `HFS_MAX_NRECS`, `hfs_issue`, `extract_child_pointers`, `record_key`, `CatalogKey`, `CatalogEntry`, `collect_catalog_entries` bumped to `pub(super)` so submodules can reach them. All 61 fsck tests still pass under their new module paths.
-  - **Not yet done:** further per-phase splits (catalog / btree). The remaining mod.rs has tighter cross-references between those two phases (catalog and B-tree share helpers like `compare_catalog_keys`, `walk_leaf_chain`, `repair_btree_structure_leaves_only`); each additional split needs a careful pass over `pub(super)` visibility. Recommended order: btree first (the hub) then catalog. Worth doing iteratively.
+  - **Done:** `src/fs/hfs_fsck.rs` is now `src/fs/hfs_fsck/mod.rs` (3054 lines, was 6588) plus four submodules: `extents.rs` (1335 lines, the 17 `*_extents_*` functions for the extents-overflow B-tree), `mdb.rs` (260 lines, Phase 1 MDB sanity), `bitmap.rs` (698 lines, Phase 4 extent/allocation-bitmap cross-check + repairs), and `btree.rs` (1322 lines, Phase 2 catalog B-tree structure checks + repairs — `check_btree_structure`, `check_map_nodes`, `check_key_ordering`, `walk_leaf_chain`, `check_index_sibling_links`, etc., plus repairs `repair_btree_structure_leaves_only`, `repair_map_nodes`, `repair_node_bitmap`, `repair_key_ordering`, `rebuild_leaf_chain`, `rebuild_index_nodes`). The shared btree primitives `record_key`, `extract_child_pointers`, `HFS_MAX_NRECS` moved with btree.rs and extents.rs imports them via `super::btree::...`. `rebuild_index_nodes` re-exported from mod.rs as `pub(crate)` so `hfs.rs` callers (`super::hfs_fsck::rebuild_index_nodes`) keep working. `CatalogKey`, `CatalogEntry`, `collect_catalog_entries`, `hfs_issue`, `HfsFsckCode` are `pub(super)` for cross-submodule access. All 61 fsck tests still pass under their new module paths.
+  - **Not yet done:** Phase 3 (catalog consistency) split — `check_directory_structure`, `check_catalog_consistency`, `parse_catalog_record`, plus the catalog repair fns (`repair_catalog_consistency`, `insert_thread_record`, `fix_thread_record`, `repair_thread_record_lengths`, `remove_thread_record`, `remove_record_by_key`, `fix_valence_mismatches`). With Phase 2 settled the catalog split should be the cleanest of the four — it can pull btree helpers from `super::btree::...` and just owns the catalog parser plus its repairs.
 
 ### HFS+ (`src/fs/hfsplus.rs`, 2607)
 - [ ] Migrate to `btree_split_leaf_with_insert` (§1).
